+++ src/modules/index.ts
@@ -12,7 +12,6 @@ export * from "./session-validation-modules/ERC20SessionValidationModule.js"
 export * from "./sessions/abi.js"
 export * from "./sessions/erc20.js"
 export * from "./sessions/batch.js"
-export * from "./sessions/dan.js"
 export * from "./sessions/sessionSmartAccountClient.js"
 export * from "./session-storage/index.js"
 import {