+++ tests/playground/write.test.ts
@@ -1,76 +1,77 @@
-import { http, type Hex, createWalletClient } from "viem"
-import { generatePrivateKey, privateKeyToAccount } from "viem/accounts"
+import { http, type Hex, createPublicClient, createWalletClient } from "viem";
+import { privateKeyToAccount } from "viem/accounts";
+import { polygonAmoy } from "viem/chains";
 import { beforeAll, describe, expect, test } from "vitest"
-import {
-  type BiconomySmartAccountV2,
-  createSmartAccountClient
-} from "../../src/account"
-import { getConfig } from "../utils"
+import { type BiconomySmartAccountV2, createSmartAccountClient } from "../../src"
+import { getBundlerUrl, getConfig } from "../utils";
 
-describe("Playground:Write", () => {
-  // const nftAddress = "0x1758f42Af7026fBbB559Dc60EcE0De3ef81f665e"
-  // const token = "0x747A4168DB14F57871fa8cda8B5455D8C2a8e90a"
-  const {
-    chain,
-    chainId,
-    privateKey,
-    privateKeyTwo,
-    bundlerUrl,
-    paymasterUrl
-  } = getConfig()
-  const account = privateKeyToAccount(`0x${privateKey}`)
-  const accountTwo = privateKeyToAccount(`0x${privateKeyTwo}`)
 
-  let [smartAccount, smartAccountTwo]: BiconomySmartAccountV2[] = []
-  let [smartAccountAddress, smartAccountAddressTwo]: Hex[] = []
+describe("Playground", async () => {
+  const { privateKey, chain } = getConfig()
+  const customChain = chain ?? polygonAmoy;
 
-  const [walletClient, walletClientTwo, walletClientRandom] = [
-    createWalletClient({
-      account,
-      chain,
-      transport: http()
-    }),
-    createWalletClient({
-      account: accountTwo,
-      chain,
-      transport: http()
-    }),
-    createWalletClient({
-      account: privateKeyToAccount(generatePrivateKey()),
-      chain,
-      transport: http()
-    })
-  ]
-
-  beforeAll(async () => {
-    ;[smartAccount, smartAccountTwo] = await Promise.all(
-      [walletClient, walletClientTwo].map((client) =>
-        createSmartAccountClient({
-          chainId,
-          signer: client,
-          bundlerUrl,
-          paymasterUrl
-        })
-      )
+  /*
+    // Alternatively, you can use the following custom chain configuration...
+  
+    const customChain = getCustomChain(
+      "Bera",
+      80084,
+      "https://bartio.rpc.b-harvest.io",
+      "https://bartio.beratrail.io/tx"
     )
-    ;[smartAccountAddress, smartAccountAddressTwo] = await Promise.all(
-      [smartAccount, smartAccountTwo].map((account) =>
-        account.getAccountAddress()
-      )
-    )
+  
+  */
+
+  const chainId = customChain.id;
+  const bundlerUrl = getBundlerUrl(chainId);
+  const account = privateKeyToAccount(`0x${privateKey}`);
+
+  const publicClient = createPublicClient({
+    chain: customChain,
+    transport: http()
   })
 
-  test.concurrent(
-    "should quickly run a write test in the playground ",
-    async () => {
-      const addresses = await Promise.all([
-        walletClient.account.address,
-        smartAccountAddress,
-        walletClientTwo.account.address,
-        smartAccountAddressTwo
-      ])
-      expect(addresses.every(Boolean)).toBe(true)
-    },
-    30000
-  )
+  const walletClient = createWalletClient({
+    account,
+    chain: customChain,
+    transport: http()
+  })
+
+  let smartAccount: BiconomySmartAccountV2;
+  let smartAccountAddress: Hex;
+
+  beforeAll(async () => {
+    smartAccount = await createSmartAccountClient({
+      signer: walletClient,
+      bundlerUrl,
+      customChain
+    })
+
+    smartAccountAddress = await smartAccount.getAddress();
+
+    const [balance] = await smartAccount.getBalances();
+    const walletClientBalance = await publicClient.getBalance({ address: account.address });
+
+    console.log("account.address: ", account.address);
+    console.log("smartAccountAddress: ", smartAccountAddress);
+
+    if (balance.amount <= 0) console.warn("Smart account balance is zero");
+    if (walletClientBalance <= 0) console.warn("Wallet client balance is zero");
+  })
+
+
+  test("should send some native token for the configured chain", async () => {
+
+    const { wait } = await smartAccount.sendTransaction(
+      {
+        to: "0x3079B249DFDE4692D7844aA261f8cf7D927A0DA5",
+        value: BigInt(1)
+      },
+    );
+
+    const { success } = await wait();
+    expect(success).toBe("true");
+
+  }, 30000)
+
 })